задание 8.14
